requirements.txt
Flask==2.3.3
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.0.5
PyJWT==2.7.0
boto3==1.28.0
python-dotenv==1.0.0
Werkzeug==2.3.7
SQLAlchemy==2.0.20
gunicorn==20.1.0
